+++ internal/provider/storage_volume_snapshot_resource.go
@@ -186,7 +186,6 @@ func (r *StorageVolumeSnapshotResource) Create(ctx context.Context, req resource
 // Read refreshes the Terraform state with the latest data.
 func (r *StorageVolumeSnapshotResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
 	var data *StorageVolumeSnapshotResourceModel
-	log.Printf("why why")
 	// Read Terraform prior state data into the model
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
 	if resp.Diagnostics.HasError() {